[0.63] Allow native module to set JS-side QuirkSettings (#8100) (#8119)
* Allow native module to set JS-side QuirkSettings (#8100)

* Allow native module to set JS-side QuirkSettings

This adds a mechanism intented to allow forwarding feature-gate values to the RN Win32/NetUI JS platform. A list of constants are present, that an optional native module can provide to override default behavior.

This will allow internally A/B'ing the VirtualizedList refactor. Need to also ensure we have a pathway set up to ensure signal.

Note that the newly added FG value is not yet present, nor is the module which will be in Office devmain to tie to FG infrastructure.

* Change files

* This doesn't need to be public

* fix flow

* Null coalesce, since falsey is valid

* override registration

* TurboModuleRegistry isn't flow strict in 0.63?