[0.65] Enable List UTs in react-native-win32 (#8121) (#8136)
* Enable List UTs in react-native-win32

This replicates react-native-windows work to enable UTs, but for react-native-win32. We share some of the special resolvers between packages. This will need to be backported across a few different branches.

Note that this change partially overlaps with #8101 which has a matching change to extract and modify the main jest resolver. This bit will merge with the first change which needs it.

* Change files

* revert vnext overrides change