Fork VirtualizedList and deps to its own package (#8101)
* Fork VirtualizedList to its own package

This extracts the internals of the RN VirtualizedList to its own package.

**Why?**
We are wanting to make some non-trivial modifications to VirtualizedList, through multiple forks. Keeping the source for VirtualizedList self-contained and working, allows us to make the package "evergreen", and ease adoption on different forks, as updating code in the old branch is just a version bump.

This is unit tested by VirtualizedList UTs running in the RNW package. Need to override duplicated files with override tooling before merge, and enable flow-checking this package after.

* Hook into overrides and rn-win32

* Fix tests

* Hook up flow

* strict-local

* Change files

* Update metro configs

* fix lint

* add comments to metro configs