[WIP] Duplicate VirtualizedList to ExperimentalVirtualizedList
We are going to flight the move to the cell render mask data struture. This change is invasive to VirtualizedList structure, so this change splits the component into two.

They are the exact same in this change, but "ExperimentalVirtualizedList" will correspond to the component included when `enableCellRenderMask` is set in QuirkSettings.