Attempt to fix e2e-test-app intermittency (#8143) (#8194)
* Attempt to fix e2e-test-app intermittency

We have seen some test intermittency after a WinUI update, where a switch will often have a different background color, and differently rounded height/width than expected. Because of intermittency and visual differences we think this might be the result of dumping the visual tree during an animation.

There are some arhitecutral improvements we could make to E2E infrastructure to add stability in case such as this, like shallow dumping (not deeply recursing into WinUI children), better syncrhonization logic, or fuzzier logic.

In this specific example we can provide an alternate solution which works around correct contraints. The purpose of the intermittent test is to validate that "diplay: none" is propagated to a textbox. It currently does this by generating a snapshot of a wider parent component, which includes the switch the test interacts with.

This change prevents us from trying to dump the visual tree of the component where we have direct transient interactions, since it is unneeded for the test. We do also have other tests which I believe dump switches in on/off state directly.

* revert

* New snapshots

* dump the outer view

* New snapshots

Co-authored-by: Nick Gerleman <ngerlem@microsoft.com>