Adds timestamp to onKeyUp/Down events (#8006)
* Adds timestamp to onKeyUp/Down events

In order to better support perf testing metrics for events driven by
either pointer events or keyboard events (e.g., triggered by Pressable,
which can be triggered by either keyboard or pointer), this adds a
timestamp property to the native event data for `onKeyDown` and
`onKeyUp`.

Ideally this would use a native timestamp from the KeyRoutedEventArgs to
get the actual time the key event occurred, but this is a sufficient
approximation absent a XAML API to retrive this.

We use `GetTickCount64()` because that matches the behavior for the
PointerRoutedEventArgs.

Fixes #8004

* Change files

* Update keyboard-reconcile-desktop proposal

* Use the same approach for keyboard event timestamps as performance.now()