Integrate 0.65.0-rc.2
This integrates an RC build from 0.65 with the updated CLI package, where we were previously using a nightly build with a couple workarounds. This build is closer to the one that will be released.

Will remove CLI workaround for older builds in a separate PR.

Generated by: `yarn integrate-rn 0.65.0-rc.2`