Fixes crash due to TextInput flyout Proofing sub-menu
There is a well known crash in XAML Islands when a MenuFlyout or
CommandBarFlyout has a sub-menu:
https://github.com/microsoft/microsoft-ui-xaml/issues/3529. This crash
is likely due to a race condition when animating the sub-menu flyout.
This change replaces the Proofing sub-menu AppBarButton with a custom
implementation that mimics the same behavior but does not encounter the
crash.