Remove AppTheme Shim
The shim returned `any` for `addListener`,  which broke typings for non-deprecated pattern of unsubsribing from the listener via calling `remove()` on the subscription returned by `addListener`.

`AppTheme` is always provided by the UWP ReactNativeHost, and we do not special case other modules in the same way for when an NM isn't present, so I just removed the shim, since the NM is always available. TS can then infer the correct subscription type from NativeEventEmitter.