Strict Handling of Node Unhandled Promise Rejections
Fixes #7804

We do not terminate or return non-zero exit codes in CI if a script has an unexpected async exception. This is due to permissive handling of unhandled promise rejections by node. The deault changes in Node 15/16 to become stricter, but these are not yet LTS. Change node options to conform to future, stricter behavior.