Use Hyperlink for pressable virtual text
We had previously created a custom component called
HyperlinkTextViewManager that derived from VirtualTextViewManager to
enable focus on pressable text, get the correct cursor behavior for
pressable text, and to work around the issue with text hit testing that
is being addressed in #7553). Rather than adding a new component that
has no meaning on other platforms, we can use Hyperlink when virtual
text meets certain conditions.

This is really just a proposal for now, we should discuss if this change
is warranted and, if so, under what conditions the Span should be
swapped out for a Hyperlink.

Please note this diff is dependent on #7553 and #7813 as we want to use
the TouchEventHandler implementation to handle clicks in a way that they
leverage the existing bubble / capture mechanisms of React Native. Also,
the Hyperlink::Click event does not given you important details like the
pointer position (see
https://github.com/microsoft/microsoft-ui-xaml/issues/4730).

Remaining TODOs are:
1. Enable handled keyboard events on Hyperlink to flow so
`usePressability` can handle keyboard invocation events.
2. Decide under what conditions we want to swap Span with Hyperlink.
This PR uses `accessibilityRole` settings as an example, but we could
also send an `isPressable` prop.
3. Add focus-related props to virtual text, so Hyperlink can set
properties like TabIndex, TabStop, and UseSystemFocusVisuals.
One remaining bit is how to enable "handled" keyboard events (as the
`Enter` and `Space` keys are handled by Hyperlink component