Adds `isPlaying` prop to Image
Adds a prop to Image to toggle AutoPlay and invoke the imperative Play
and Stop methods on BitmapImage to stop GIFs from animating. This is a
useful feature for apps that wish to stop GIFs when the app goes out of
focus or when an Image component leaves a ScrollView view port.

I believe this prop can and should eventually be upstreamed into React
Native, adding here for now as we only have an immediate need for
Windows.

Fixes #8341