Work around XAML Islands issues with TextCommandBarFlyout (#8249)
* Work around XAML Islands issue with TextCommandBarFlyout

As reported in
https://github.com/microsoft/microsoft-ui-xaml/issues/5341, the
singleton TextCommandBarFlyout for the XAML TextBox and TextBlock
components are limited to one window. When the flyout opens on a
different window after already being opened on another, the first window
the flyout was opened on gets focus, and the flyout is immediately
closed. This change creates a TextCommandBarFlyout per TextInput native
view and per selection TextBlock native view.

* Change files

* Adds helper to share logic between TextInputViewManager and TextViewManager

* Adds fix for microsoft-ui-xaml#3529 for proofing menu

* Clear flyout property settings when TextBlock is no longer selectable

* Ensure multi-window flyout issue is also applied to PasswordBox