Add root tag to ShadowNode for resolving correct XamlRoot in FlyoutViewManager (#8327)
* Add root tag to CreateView in ViewManagerBase

There are a number of places where a single root view assumption exists
in the app. Currently, there is no way to know which root view a Flyout
will be shown on, which can be problematic when each root is on a
different window. This is the first step to supporting multiple root
views.

Towards #8250

* Change files

* Adds API to get specific XamlRoot to NativeUIManager

* Get specific XamlRoot for FlyoutViewManager

* Addresses review feedback