Clear tintColor when prop set to null
Sets color to Transparent when prop is explicitly set to null. Setting
the color to Transparent will set the A value to 0, which will allow us
to switch from a ReactImageBrush to a standard bitmap. It will also
remove the tint effect from the CompositionEffect for images that use
either repeat resizeMode or blurRadius.