Suppresses auto-focus under certain conditions
In multi-window XAML Islands apps that share a single UI thread for each
window, focusing a view on a blurred window causes that window to get
focus. Generally speaking, focusing a window should be explicit, and
it's likely very rare that you would want a view getting focus to cause
the entire window to get focus.

This manifests as problematic in two ways:
1. If you have a flyout shown and the flyout gets dismissed because you
switch to another window, the parent window for the flyout would steal
focus again when react-native-windows attempts to focus the target
element.
2. If you re-render a view with a TextInput with `autoFocus={true}` and
the parent window for that TextInput is blurred, the window will steal
focus.

This change forces the user to first focus the window and update the
active XamlRoot via the `XamlUIService::SetXamlRoot` API before a view
in the window can get focus.