Remove or Migrate Usages of Picker
The [6/23 RN integration](https://github.com/microsoft/react-native-windows/pull/8357) removes Picker. We still have a few usages of it in custom RNTester examples or the like. This change removes all usages of it from our own code, either removing where inapplicable, or using the reccomended community module instead. This gives us a bit of real coverage of a module in our repo, but we only require it for the RNW RNTester fork instead of including it in RNW itself.

This exposed an issue where we broke usageof RNW with existing community modules (see https://github.com/microsoft/react-native-windows/issues/8430). We workaround this with patch-package for now, which also has us enabling yarn scripts again so we can hook postinstall.

Playground and playground-win32 are both updated to support autolinking, and it is tested that community module picker controls work correctly in them.