Implement native inverted behaviors for ScrollView
This PR modifies ScrollViewManager and VirtualizedList to work around
the limitations of the RN core approach to list inversion. Specifically,
the RN core approach to list inversion uses scroll axis transforms to
flip the list content. While this creates the visual appearance of an
inverted list, scrolling the list via keyboard or mouse wheel results in
inverted scroll direction.

To accomplish native inverted scroll behaviors, we focused on four
expected behaviors of inverted lists:
1. When content loads "above" the view port in an inverted list, the
visual content must remain anchored (as if the content renders below it
in a non-inverted list).
2. When scrolled to the "start" of an inverted list (in absolute terms,
the bottom of the scroll view), content appended to the start of the
list must scroll into view synchronously (i.e., no delay between content
rendering and view port changing).
3. When content renders on screen, it must render from bottom to top,
so render passes that take multiple frames do not appear to scroll to
the start of the list.
4. When imperatively scrolling to the "start" of the content, we must
always scroll to the latest content, even if the content rendered after
the scroll-to-start animation already began.

For 1., we leverage the XAML `CanBeScrollAnchor` property on each
top-level item in the list view. While this is an imperfect solution
(re-rendering of this content while in the view port can result in view
port shifts as new content renders above), it is a good trade-off of
performance and functionality.

For 2., we leverage the XAML `HorizontalAnchorRatio` and
`VerticalAnchorRatio` properties. XAML has a special case for inverted
lists when setting these property values to `1.0`. It instructs XAML to
synchronously scroll to and render new content when scrolled to the
bottom edge of the ScrollViewer.

For 3., we leverage Yoga's implementation of `flexDirection: column-reverse` and `flexDirection: row-reverse` to ensure content is rendered from bottom to top.

For 4., we implemented `ScrollViewViewChanger` to continuously check if
the target scroll offset has changed since starting an animated
scroll-to-end operation. If the target scroll offset no longer matches
the scrollable extent of the ScrollViewer, we update the target offset
by calling `ChangeView` again.

Fixes #4098