Initial implementation of TextVisitor
We have 3 different features that require a DFS traversal of the text
tree:
1. The textTransform algorithm needs to recurse the undefined paths of
the Text sub-tree where the prop value changes to update.
2. The backgroundColor algorithm needs to recurse the entire text tree
to re-calculate highlighters any time the text changes or
backgroundColor / foregroundColor props change within the tree.
3. The WIP hit test algorithm needs to recurse the text tree to identify
which span is pressed.

Additionally, we have 4 "upward" traversal algorithms:
1. When a RCTRawText changes it's text prop, we need to send an
accessibility notification
2. When a RCTRawText changes, we need to re-apply text transforms.
3. When a background or foregroundColor changes, we need to notify each
parent that it may require a highlighter for an optimization that skips
calculating highlighters for text trees without foreground or
background colors.
4. In the WIP hit test algorithm, we need to notify parent nodes that a
nested Text component is pressable for the optimization that selectively
skips traversal of non-pressable sub-trees

Each of these algorithms follow the same pattern, so reduce the
potential for bugs, I'd like to consolidate these algorithms to use a
visitor pattern.