Fixes issue with inherited text transforms
Previously, if the child of a virtual text node with undefined
textTransform was appended, and that virtual text node had a parent with
a defined textTransform, the textTransform value would not inherit
properly. Similarly, if a virtual text node switched from defined to
undefined for textTransform prop, it would not inherit the correct value
and update children accordingly either.

Fixes #8460