Set Shared WindowsTargetPlatformVersion in External Property Sheets
Fixes #8443

External property sheets are managed by the react-native-windows npm package instead of the template. This means that a single version defined by the react-native-windows npm package will be included for any community modules.

Set WindowsSdkVersion there instead of in the template so that it stays in sync with the host app. The current way of overriding SDK version in ExperimentalFeatures.props will still work because it is imported first.