Generalizes hit testing so all view managers can participate
One of the comments in #7553
(https://github.com/microsoft/react-native-windows/pull/7553#discussion_r640427024)
was that a community module view manager would not be able to implement
hit testing the way we implemented it for TextViewManager. This change
generalizes hit testing so any view manager can participate. We don't
(yet) have a use case for community modules to participate in hit
testing, but presumably when we do, we can add an
`IViewManagerWithHitTest` interface that allows community modules to
participate.