Adds hook for view managers to capture pointer events
Currently, there is no way for ViewManagers to capture the pointer to
perform it's own pointer event handling. This change introduces a new
interface, `IViewManagerWithPointerEvents`, so third party modules can
be notified when the ReactRootView receives pointer events. The behavior
is as follows:

If `IViewManagerWithPointerEvents` returns `true` in `OnPointerPressed`,
`TouchEventHandler` will not capture or start tracking the pointer for
React Native touch events.

If `IViewManagerWithPointerEvents` returns `true` in `OnPointerMoved`,
`TouchEventHandler` will release capture of the pointer.

If `IViewManagerWithPointerEvents` returns `true` in
`OnPointerReleased`, `TouchEventHandler` will send a cancellation event
to React Native rather than an end event.