Rename node-rnw-rpc to @react-native-windows/automation-channel
The node-rnw-rpc package is used to enable test runners to interact with an RNW application, and is used in e2e-test-app and integration-test-app. It enables scnearios, such as sending the XAML tree over the wire to Jest.

There has been some interest in using this external to the RNW repo. This rename prepares for moving some of the common automation infra in the test apps to the native module that may be shared.