Use `Target` property on ReactPointerEventArgs
Rather than maintaining two places where we walk the view tree, this
change replaces the existing TagFromOriginalSource method with one that
leans on the new IViewManagerWithPointerEvents. This will allow us to
decouple TouchEventHandler to implement things like `pointerEvents` for
`box-none` and `box-only` and also text hit testing in future PRs.