@react-native-windows/automation
This combines and documents some packages to allow external applications to leverage the work we did to allow UI testing. The various READMEs in this PR lay out what the user-flow and architecture looks like, but there is still some work around extracting IPC/tree dumping to be externally usable.