ReactUwpTestApp -> RNTesterApp
The name ReactUwpTestApp is an artifact of when `ReactUwp` still existed, and having a specific test bundle. Now the app is just a frame to load RNTester, plus side native test code that will soon be extracted.

This PR renames ReactUwpTestApp to a more accurate RNTesterApp. It also removes some stale scripts that have been unused for more than a year.