Fix logic to avoid launching packager (#8541)
* Fix logic to avoid launching packager

We do not want to launch the packager by default in release builds because the application is already bundled. Logic is present to avoid launching, conditioning on release mode, or if  ``--packager`` is specified.

Commander has some special handling around negation parameters. For a positive parameter, like `--foo`, the option will be set to undefined ifa default not present. If it is `--no-foo`, then `foo` will be set to true instead of undefined.

This effectively means there is not a way to test for `--packager` vs the absence of `--no-packager`, so the previous logic always assumed the user was explicitly asking for a packager.

This change moves the logic to instead launch the packer when the following conditions are met:
1. `--no-packager` is not set
2. `--no-launch` is not set
3. `--release` is not set

* Change files

* Remove default