Fix autolinking to respect WinUI2xVersion or WinUI3Version set in ExperimentalFeatures.props
This changes autolinking to look for those properties in
ExperimentalFeatures.props and if not found, then fallback to the values
in WinUI.props.

Note that while this "fixes" autolinking for playground win32
(unpackaged) it "breaks" autolinking for playground win32 (packaged) and
playground UWP. So regardless of this change, autolinking should remain
disabled when building any playground.

Tests have also been added to verfiy this change in logic.

Closes #8431