Adds functionality for box-only and box-none via OnPointerEvent callback
This change checks if the React node has `pointerEvents="box-only"` or `pointerEvents="box-none"` set. If the node is `box-only`, it updates the `ReactPointerEventArgs.Target` view to itself (preventing children from becoming the hit test target for the React pointer event. If the node is `box-none` and the current `ReactPointerEventArgs.Target` value is the current view, it sets the target to `null`, forcing the `TouchEventHandler` to choose an ancestor as the hit target.

Fixes #8496