Fixed CLI options defaults (#8629)
Commander (the library used by the RN CLI to define/parse CLI args) does
not support setting defaults for boolean flags. All such options are
resolved to either true (if present) or undefined (if specified in
the --no-flag format and present).

This PR was extracted from PR #8253, and cleans up the CLI options to
not specify defaults for boolean flags, as they are ignored by Commander
and are therefore potentially misleading. The RunWindowsOptions and
AutoLinkOptions interfaces have been updated to reflect these changes,
and tests were added to make sure we don't set useless defaults in the
future.

This PR also fixes the formatting of autolink.test.ts, which did not
have the proper copyright header and therefore was not being checked by
eslint.

Closes #8630