fix(deploy): broken projectName override for deployToDevice
when `projectName` is specified in the `windows` config block and
more than one `AppxManifest.xml` is found, `projectName` should be
used to disambiguate as intended. this works when called from
`deployToDesktop`, but `deployToDevice` does not currently pass the
overridden `projectName` to `getAppxManifestPath`.

add a `projectName` parameter to `getAppxManifest` and modify
`getAppxManifest` to pass the argument along to
`getAppxManifestPath` instead of `undefined`.