feat(deploy): allow overriding build output dir
some projects override the default build output directory by
specifying `OutDir` in their project files or imported props. as a
result the `AppxManifest.xml` and `AppPackages` directory are not
present under the default `windows` path and aren't found by the
existing glob patterns in `getAppxManifestPath` and
`getAppPackage`. this causes `react-native run-windows` to fail
to deploy the app for these projects.

add an optional `projectOutputDir` field to `WindowsProjectConfig.Project`
to allow users to explicitly provide the path to build output where
the `AppxManifest.xml` and `AppPackages` directory are placed.

ideally the `$(AppxPackageDir)` MSBuild property would be used to
determine the`AppPackages` directory and `$(OutDir)` MSBuild
property would be used to determine the path to `AppxManifest.xml`,
but these can be set in any imported props/targets file and there
is currently no way to get the effective runtime value.