Ensures IsHitTestVisible is set on correct view (#8634)
* Ensures IsHitTestVisible is set on correct view

When someone sets `pointerEvents="none"` on a View component, we
currently set `IsHitTestVisible` to false on the ViewPanel instance. If
the user has done something that forces the ViewPanel to be replaced by
a ViewControl, we will not have set `IsHitTestVisible` on the
ViewControl and will not also transfer that property if
`pointerEvents="none"` is set before the transfer occurs. This change
ensures IsHitTestVisible is always set on the correct view.

Fixes #8531

* Change files