Enable codesigning in main, enable symbol packages for everyone
This PR changes the publish pipeline to start codesigning daily main
release builds, so they can start to be published to Nuget.org. This
PR also enables symbol packages for all release builds.

This is a continuation of the work started in PR #8626 for issue #6011.

This also sets the groundwork to be able to resolve issue #8529.