[0.63] Fix ARM64 layout issues in Office (#8676)
* Change the definitino of YGUndefined from NAN (contra the comment in
YGValue, MSVC does define NAN) to __builtin_nanf("0"), which generates
a quiet NaN. The problem with NAN is that corecrt_math.h defines it as:

Which *seems* fine! However, because we are compiling with /fp:precise
(see: https://github.com/microsoft/react-native-windows/issues/4122 for
more details), that multiplication actually happens at runtime. And,
unfortunately, we're sometimes seeing MXCSR set to round down (i.e., the
_MM_ROUND_DOWN flag is set)), which, combined with the SSE instructions
that happen to run here (specifically, cvtpd2ps), cause NAN, and thus
YGUndefined, to yield a value of 0! What happens is that cvtpd2ps, when
rounding down, converts 1#.INF in double-precision not to 1#.INF in
single-precision, but instead to FLT_MAX!!! And, of course, FLT_MAX*0 is
0, not qNaN. Oops.

Anyways, using __builting_nanf("0") just loads 0x7FC0000 into xmm0,
rather than doing any math, so, even with /fp:precise + bad rounding
modesm, YGUndefined is, in fact, a quiet NaN.

* Change files

* Add YGValue.h to overrides.json