Remove RNW peerDep from automation-commands package (#8573)
* Remove RNW peerDep from automation-commands package

Fixes #8569

When RNW has a prerlease bump, Beachball will overwrite set peerDependencies in dependent packages. `@react-native-windows/autoation-channel` is set to contain a native module, so it allowed RNW 0.64 or greater, or canary versions. Beachball instead changed it to only support our latest canary, then also made packages and dependencies prerelease, not to be picked up automatically in the future.

This behavior is very much not what we want. I did a quick scan of beachball internal options to see if we could customize it, and didn't see anything. Removing the peerDep for now, to get more sane behavior, This will break PNPM usage, but that will affect less folks, and can be addressed when it becomes an issue.

* Change files