Refines algorithm for hit testing Text (#7553)
This algorithm implements multiline text hit testing for virtual text nodes. Previously, hit testing was limited to single line text. This algorithm adds an `OnPointerEvent` override to TextViewManager and conducts a "normal" case O(1) hit test on each Run that is a descendant of a pressable virtual text node. The algorithm is worst case O(log(n)) when the Run contains RTL text in an LTR context or LTR text in an RTL context, however the character simply looks for the left-most RTL character on the first line and the right-most character on the last line (for RTL in LTR flow) to identify the bounding boxes for the first and last lines (it does not search over the entire run).

There is a known limitation to this algorithm, as the XAML APIs do not include character width information. So, for example, the final character of the first line of RTL text in LTR flow for multiline or bi-directional text will fail hit testing (filed as an issue in #7792).

Fixes #5151