Fixes crash when animating Opacity on custom view with TranslateTransform
If you create a custom view that sets the UIElement::RenderTransform()
property to TranslateTransform, then attempt to use an Opacity animation
from NativeAnimatedModule, the app will crash because it attempts to set
the RotationAxis property, even if we are not applying a rotation
animation.

This change ensures that the RotationAxis property is only set when
we're attempting to animate rotation.