Set Automation Peer for ReactImage
We don't currently create a UIA node for images. FrameworkElementViewManager (which ImageViewManager dervives from) converts RN a11y props to attached properties on the element. ReactImage, unlike ViewPanel, does not however customize Grid's OnCreateAutomationPeer.

This hooks ReactImage into using ViewPanel's automation peer, which can understand the set accessibilityRole.

After this change, the image shows up in UIA with localizedControlType of `image`. It will also now propagate `accessibilityLabel`, and `testID`. By default the image is **not accessible**, matching the behavior of other platforms. This translates to being added to the raw tree by default, vs the content tree when `accessibile` is set to true.

It would be good to query automation properties in UI tests, but we are blocked on #8801 to do that.