Template App WACK Validation and YAML Build Matrix Refactoring
We currently only run WACK validation for the playground app.

#8812 discovered that we fail WACK validation for default apps, that do not copy Hermes DLL into AppX.

In preparation for a fix, this change enables WACK validation for all template apps we build in release mode. This is done separately from the fix to ease backporting. We temporarily set `continueOnError` to allow us to merge this before the fix.

Template sharing ran into issues with runtime matrix expansion of init test job. I ended up moving all of our jobs to template-time build matrices to allow simplifying a lot of our logic, and to allow typed parameters in the build matrix. This is the bulk of this PR.