Fixes bug introduced by e7735a8 (#8820)
* Fixes bug introduced by e7735a8

In order to properly set the `pageX` and `pageY` values in
`TouchEventHandler`, we need to ensure the root view is set, which is
used to calculate these properties. By default, the root view is not
provided to the `TouchEventHandler` so we should treat the supplied
`xamlView` argument as the root view.

* Change files