Remove Proofing menu item when spellCheck disabled
There is a crash workaround for XAML Islands that replaces the Proofing
menu item with a custom AppBarButton that does not crash. This
workaround was failing to remove the Proofing item after the spellCheck
setting was disabled. This change ensures that the Proofing item is
removed when it is no longer needed.

Fixes #8832