Conditionally fires `onDismiss` when receiving `ContentDialogResult::None`
When the ContentDialog returns with a result of `None`, and there is no
close button / neutral button option presented to the user, this change
fires an `onDismiss` instead of attempting to fire the
`buttonNeutral.onPress` callback, which is very unlikely to be
registered.

It is very unlikely that RNW apps have leaned on this, but if someone is
depending on the implementation detail of the "neutral" callback being
called without presenting the button (by leaving the text for the
neutral button option null), then this is a minor breaking change.

Fixes #8865