Stamp RNW Version into MSBuild Props (#8772)
* Stamp RNW Version into MSBuild Props

This change allows community modules to query against the RNW version used at compile-time. This creates a foundation for conditional compilation against new features/APIs/quirks.

In order to calculate the new version number, existing DLL stamping code is run after the NPM package is published. Because we want to insert the version into the NPM package, we need to hook-in earlier. We use beachball prePublish hook as a mechanism to do this,. For clarity, we also commit the generated version to the repo, alongside the change to update package.json.

Using beachball hooks requires us moving our root beachball config from our packge.json to a JS file. We previously mutated beachball args in the package.json, so we need to rewrite some of the code which generates beachball config.

Props source of truth is a generated PropertySheet, exposed to internal and external projects. Preprocessor definitions mirroring the props are exposed by default to C++ apps/libs via external property sheets in the npm package.

Stamping is done via a JS script and mustache templates (ala generator). This should make it easy to stamp versions into more file formats in the future (E.g. for runtime queries in C++/C#).

* ordering

* Add macros

* Update beachball.config.js

* yarn format

* fiz

* Update vnext/template/cpp-lib/src/pch.h

* formatting and git

* fix comparison

* Use existing commit message for familiarity

* Update TypeScript in package.json

* parens and subsequent `yarn format`

* inverted quotes