Ensure tail spacer is not used as a scroll anchor
We previously stopped development on the native inverted approach for
Windows that would use anchoring and flexDirection to implement a truly
reversed list (as opposed to a list flipped via transform) due to a bug
where the anchoring would cause the view to "ride" to the top of the
ScrollViewer when you scrolled into the virtualized tail spacer.

This change introduces a new View prop called `overflowAnchor` and uses
this prop to ensure that the tail spacer can never be anchored.