Download less artifacts in NuGet Init Test
One of the tests that tends to finish last is the init test using NuGet package. This test relies on the build in a first stage, then in its own, it locally creates the NuGet package for testing. Downloading artifacts currently takes over a minute.

This PR aims to decrease total PR time by not downloading build artifacts for slices that are not needed.