Set default AppxBundlePlatforms (#9006)
* Set default AppxBundlePlatforms

The `AppxBundlePlatforms` property determines which platforms are added to the generated AppX. AppX target logic to detemine a default for this is a bit suspect, and will force a rebuild of the project in any architectures not already built.

This change makes sure we only build a single platform at a time for our debug builds, and makes sure we build ARM64 by default for release builds.

To keep CI parallel, I've added an env setup step to only build a single platform at a time in PRs. I did not change CI logic, in order to emulate a more stock configuration. This means each release build in CI will however build all three platforms I think, and this may require some adjustments. Needed to clean up our CI environment init code a bit to apply this eveywhere it should go.

* Change files

* yaml parse fix

* yaml

* single quotes

* Missing path

* typo

* Move reunion to typesafe matrix

* fix

* fix

* prepare-build-env in UniversalTest

* Needs JS env?

* typo

* Remove duplicate steps