Add low resource jobs
As we start to take advantage of our beefier hardware running our CI infrastructure, we lose signal on build reliability of lower-end hardware.

This validation ensures that we can build debug and release on Azure Standard_DS2_v2. This SKU is chosen as it is used by both Azure Pipelines and GitHub hosted agents.