Fixes default background for mouse events with border
When `onMouseEnter` or `onMouseLeave` is set, react-native-windows sets
a transparent background by default on a View if it has no other
background. Since we were setting the `Panel::Background` property
directly instead of using the `ViewPanel::ViewBackground` property, the
transparent background could be cleared out after being set. This change
ensures that the transparent background is set even if a border is
applied.

Fixes #9034