Per-Slice NuGet artifacts in CI (#9019)
* Depends on #9017 to be merged first to avoid failures of the CI pipeline

We currently aggregate the various universal flavors into a UniversalBuild job, which must be completed before NuGet package based jobs. This means we may be waiting to start debug-flavor NuGet CLI tests until the longer release is also finished. It also adds a bit of extra latency.

This removes the step, and instead relies on only the flavors needed. In doing so we also move NuGet header layout to package creation time, instead of depending on a single job to also upload header artifacts. We no longer rely on MSBuild to patch react, so this can now be done without building.

This also partitions pipeline artifacts per-slice, to let NuGet init tests only download a single platform.

* remove changefile

* Change files

* PublishToolDuringBuild non-x64

* Add Microsoft.ReactNative.Managed.CodeGen.UnitTests

* MSBuild condition fix

* Typed slices with isolated artifacts

* yaml

* yaml

* convertToJson

* ARM casing in matrix names

* Update UniversalTest download path

* yaml

* Move PackDesktop to agent pool