Segregate artifacts in publish (#9043)
* Segregate artifacts in publish

I switched from publishing build artifacts to publishing pipeline artifacts. This seems to change behavior to no longer allow overlapping artifact names, where the previous allowed them so long as there were not overlapping files within the artifact.

Artifacts are already partitioned in PR, but are still shared in publish. This change makes it so Desktop, MSRN, and Reunion are separate artifacts in publish as well.

This also reminded me that Reunion has a different version of MSRN, which isn't an issue with segregated NuGet creation in PR, but needs special attention in CI.

* Explicit slices

* yaml