Automatically Use Large PCH on Machines With >7GB per Thread
Large PCH saw issues with previous machines, which were 3.5GB per thread. New machines have 8GB per thread, and do not see issues. Enable using large PCH in a known good scenario, so that customers will see the same things we are doing in our CI.

We have tests now for both scenarios, so we should see coverage so long as we continue to test with lower end agents.