Removes unnecessary prop settings on VM base class
Removes the explicit setting of layout prop values on
FrameworkElementViewManager for:
- width
- height
- minWidth
- minHeight
- maxWidth
- maxHeight

Setting these values can cause the bug described in #9050 due to the
special treatment of numeric values for these props.

These props should not be needed at all. If a view is self-measuring,
Yoga hands the parameters of layout via the self-measuring callback, and
at any rate the width and height props are cleared before measuring:
https://github.com/microsoft/react-native-windows/blob/8adef58a52ed3400ba288fa4299e73fb8eef8384/vnext/Microsoft.ReactNative/Views/ViewManagerBase.cpp#L70-L71

And the `minWidth` / `maxWidth` props will be accounted for in the Yoga
layout calculation when setting the width/height bounds and measurement
modes.

Fixes #9050