Remove extra definition of ReactNativeWindowsDir
ReactNativeWindowDir is normally defined in the top-level app/lib project, using logic that emulates node resolution. There is a second definition in external property sheets, that may be defined if folks have a custom project without defining ReactNativeWindowsDir. This falls down, since we use ReactNativeWindowsDir before it is defined in the file.

This change removes the second copy of ReactNativeWindowsDir, and adds an error to try to catch folks homebrewing their own project.