Fixes #9063 (#9064)
We use RoslynCodeTaskFactory to compile inline C# in our MSBuild scripts and execute it with .Net Core. When running under ARM64, DotNetSdk_LibraryPath_arm64 = C:\Program Files (x86)\Windows Kits\NETFXSDK\4.8\lib\um\arm64 is added to the csc library path, but is unused, and not actually present by default. This generates compiler warnings, failing the build with MsBuildWarnAsErrorArgument, despite creating a succesful assembly. This also affects our ARM64 NuGet packages, which run inline C# as part of their targets. We were not testing these in CI though.

This change alters the environment to supress the warning, allowing our CI tasks with /warnaserror to pass.

Validated locally that the warning is no longer present when building the Microsoft.ReactNative project on ARM64.