Ensure native animations fire completion callbacks
The AnimationDrivers were being stored as unique_ptrs, but internally
use a weak_ptr to handle completion callbacks. When attempting to
convert the weak_ptr to a shared_ptr, the conversion would always return
an empty pointer because the constructor was not called from a
shared_ptr.

This changes the NativeAnimatedNodeManager to hold shared_ptrs to
AnimationDrivers to ensure that callbacks fire when the composition
scoped batch completes for the animation.

Fixes #9157