Telemetry Refactor (#8253)
This PR refactors our RNW CLI telemetry code to be easier to consume, align with our requirements, and add a suite of unit and E2E tests.

**Note:** This PR does not *enable* telemetry collection yet - we still need to (re)-setup the service side to actually receive data, test the E2E pipeline and fix any bugs.

Closes #7312
Closes #7903

## Overview

All telemetry-related operations are driven by the `Telemetry` class, the `CodedError` type, and a small handful of helper functions exported from the `@react-native-windows/telemetry` package. Consumers of this package (`@react-native-windows/cli` and `react-native-windows-init`) are provided this abstraction layer and are therefore no longer expected to understand the intricacies of the underlying telemetry service (Application Insights).

CLI commands are now generally restructured to follow this pattern:

```ts
async function cliCommand(args) {
  // Telemetry setup
  Telemetry.setup();
  Telemetry.startCommand(...);

  let cliCommandError: Error | undefined;
  try {
    await cliCommandInternal(args);
  } catch (error) {
    // Report any error thrown
    Telemetry.trackException(error);
    cliCommandError= error;
  }

  // Telemetry cleanup
  TelemetryTest.endCommand(...);
}
```

With this setup, none of the code within internal function (`cliCommandInternal` here) has any notion of or connection to the telemetry system, other than throwing `CodedError` whenever it hits a known error in our code that we wish to capture. (Note that none of the functionality of any of our commands were changed in this PR, as they were already throwing `CodedError` s with appropriate extra data, where appropriate). Any unknown errors are also caught and recorded by the Telemetry.

In the end, we expect up to two events from every CLI command: `RNWCLI.Command` and potentially one error event: `RNWCLI.CodedError`. The `RNWCLI.Command` event is fired in the call to `endCommand` and contains everything we want to know from a run of our commands, success or failure. If an error was thrown, then we'll also have fired an additional event with the call to `trackException`, including error codes, stack traces, etc. The two events can be correlated to one another via the `sessionId` property.

Every effort has been made to make sure that we can easily extend and update our CLI commands *without* having to update the telemetry code. For example, the work to gather the command line arguments/options passed, to gather project info and environment setup, are all set up in a standardized way and sanitize such data of PII.