Fixes onMouseEnter/Leave to fire when pointer down
We fire onMouseEnter/Leave from `UpdatePointersInViews`, which
historically was suppressed on `PointerMove` because we wanted to
prevent `onMouseMove` from firing while the pointer was down. Since we
no longer have `onMouseMove` behavior in react-native-windows, this
change ensures that `onMouseEnter` / `onMouseLeave` (and thus hover
states) update regardless of the pointer being up or down.

Fixes #9171