Add environment variable overrides to specify telemetry setup string and proxy
We need a way to be able to specify different setup strings and proxy servers for the telemetry. We can do this in code, but that's not always an option, for example in the case of `react-native-windows-init` which is called by using npx.

Adding these environemnt variable overrides will help us test the telemetry system before (and after) we go live and turn it on.

Example:

```cmd
set RNW_TELEMETRY_SETUP=12345678-abcd-abcd-abcd-1234567890ab
set RNW_TELEMETRY_PROXY=http://localhost:8888
```

Closes #9166