Fixes issue with restarting animations
Animated Composition values cannot be queried until the animation
completes (generally, 1 full frame callback after the animation is
stopped). The value is also ready to be queried after a scoped batch
completion event fires.

This change introduces deferred animation starts. If we attempt to start
an animation on a ValueAnimatedNode where an active animation is already
running, we wait for the `Completed` callback on that animations scoped
batch to fire before starting the next animation. The `Completed`
callback will fire either after the animation has run to completion or
roughly one animation frame after `StopAnimation` is called.

Doing this ensures that we call `FlattenOffset` so the starting value
for the ValueAnimatedNode is correct when we start the animation next.

Fixes #8419